+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Changelog
 
+### Version 0.0.1a6 (25-Mar-2026)
+- Updated `startProgressDialog()` to show the resolved operation titles that will actually run in interactive modes
+- Wait for the background progress dialog to close before continuing
+- Suppressed `swiftDialog` stderr for captured JSON dialogs
+
 ### Version 0.0.1a5 (18-Mar-2026)
 - Enabled moveable and minimizable window for `startProgressDialog()`